+++ CHANGELOG.md
@@ -9,3 +9,11 @@ Show request latency in output.
 ## Version 1.2.0
 
 Show response headers in output.
+
+## Version 1.3.0
+
+- Show error as popup, not as new text document
+- Provide a "Run HTTP" Task
+- Add support for request headers
+- Add support for request body
+- Separate multiple requests with a `###` on a separate line